Moved `MatchClosure` Expression from tests to library. There is no Generics at PHP and interface requires implementation's method signature to be completely equal to the one at interface so had to use `instanceof` here.
